requirements.txt
Django==3.1.2
django-filter==2.4.0
django-property-filter==1.0.0
django-widget-tweaks==1.4.8